// web/lib/api/generated/model/ordersListTab.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */

export type OrdersListTab = typeof OrdersListTab[keyof typeof OrdersListTab];


export const OrdersListTab = {
  backorders: 'backorders',
  completed: 'completed',
  in_progress: 'in_progress',
  new: 'new',
  on_hold: 'on_hold',
} as const;